fix(cli): split outboundUrlGuard's DB helpers so setup-opencode packages cleanly (#7682)
