+++ tests/unit/webhook-metadata-guard-3269.test.ts
@@ -14,8 +14,11 @@ import path from "node:path";
 
 process.env.DATA_DIR = fs.mkdtempSync(path.join(os.tmpdir(), "omni-wh-meta-3269-"));
 
-const { parseAndValidateWebhookUrl, isCloudMetadataHost, OutboundUrlGuardError } = await import(
+const { isCloudMetadataHost, OutboundUrlGuardError } = await import(
   "../../src/shared/network/outboundUrlGuard.ts"
+);
+const { parseAndValidateWebhookUrl } = await import(
+  "../../src/shared/network/outboundUrlGuardPolicy.ts"
 );
 const { resetDbInstance } = await import("../../src/lib/db/core.ts");
