+++ tests/unit/webhook-private-optin-3269.test.ts
@@ -14,8 +14,9 @@ import path from "node:path";
 
 process.env.DATA_DIR = fs.mkdtempSync(path.join(os.tmpdir(), "omni-wh-3269-"));
 
-const { parseAndValidateWebhookUrl, OutboundUrlGuardError } = await import(
-  "../../src/shared/network/outboundUrlGuard.ts"
+const { OutboundUrlGuardError } = await import("../../src/shared/network/outboundUrlGuard.ts");
+const { parseAndValidateWebhookUrl } = await import(
+  "../../src/shared/network/outboundUrlGuardPolicy.ts"
 );
 const { resetDbInstance } = await import("../../src/lib/db/core.ts");
